Move ImplicitVariantReturnTypeInspectionTests for QuickFixes into relevant files under RD.QuickFixes
